refactor: rename config.py to user_config.py to avoid confict with config/ command directory
