+++ scaffoldr/template_handler.py
@@ -4,7 +4,7 @@ from tomllib import TOMLDecodeError
 from tomllib import load as toml_load
 from typing import Any, TypedDict
 
-from scaffoldr.config import USER_DEFINED_TEMPLATES_PATH
+from scaffoldr.user_config import USER_DEFINED_TEMPLATES_PATH
 from scaffoldr.exceptions import TemplateError
 
 BUILTIN_TEMPLATES_PATH = Path(
